DEV: Update HTML Coverage Recipe (#64)
Change recipe in Makefile for coverage to use pytest-cov instead of
coverage.

Change recipe in Makefile for coverage to use pytest-cov instead of coverage. Pytest-cov supports multiprocessing and a few extra bits. Also added pytest-cov to dev requirements so that it will not throw and error when using the recipe. Also for consistency across testing (travis / tox is using pytest-cov and the env is slightly different when using coverage).

To generate test coverage locally run the make recipe:

```bash
$ make coverage
```

This will run the tests and collect coverage before opening the index in a web browser:

![image](https://user-images.githubusercontent.com/4615004/102013978-5ef6b480-3d4b-11eb-8ac8-6db2f340179a.png)

Clicking through to a file will show line-by-line coverage:

![image](https://user-images.githubusercontent.com/4615004/102013988-72a21b00-3d4b-11eb-9efe-d0abfa836702.png)

Co-authored by <John-P>