ENH: Enhance stain normalisation luminosity stand (#65)
To be able to perform most of stain normalization algorithms, we need to extract stain matrix from the stained tissue in the image. In other words, the stain matrix should encode only the color information in the tissue region of the image. However, given a target image, that image might include background region as well which can affect the stain matrix and hence the final stain normalization output. Therefore, the first preprocessing step in the stain normalization would be tissue segmentation. 

As we usually use small target image which mostly contain the tissue region, using adaptive thresholding algorithms (like Otsu) would results to imperfect segmentation of target image, for example separating the stromal region from nuclei regions. Therefore, we prefer to use fixed thresholds to separate between background and foreground (tissue) in the image. However, due to the variation in the intensity levels of images, choosing a single threshold may not work for all cases. To overcome this challenge, we propose to first standardize the illumination of the input image and then apply the threshold.

The proposed standardizing method is image contrast enhancement algorithm by linearly adjusting the image intensities. Here, we first find the highest and lowest 2% percentile of the image intensity values, and linearly change all the pixel intensities so that those high and low percentile values become 255 and 0:
`img_out = skimage.exposure.rescale_intensity(img_out, in_range=(p_low, p_high), out_range=(0., 255.))`

This new implementation changes both high and low intensity values in the image and therefore guarantees to achieve a better contrast in the output image in comparison to other method (in stainTool) which adjusts only high percentile of image intensities and would results to a low-contrast image. In the following figure, you can see the effect of each preprocessing method. 

![luminosity](https://user-images.githubusercontent.com/74412979/102600400-aba90980-4116-11eb-8b19-e6063a4639f0.PNG)

As illustrated in the above figure, the luminosity component of the image after stainTool adjustment become very low-contrast that even choosing a high threshold of 0.98 would results to segmenting both background and tissue region (and may miss some tissue parts). On the other hand, in the proposed contrast enhancement method, one can choose a reliable threshold of 0.9 and be sure that only tissue regions are being segmented.

Co-authored-by: Mostafa Jahanifar <mostafajahanifar>
Co-authored-by: Simon Graham <simongraham>
Co-authored-by: Shan E Ahmed Raza <shaneahmed>