NEW: Add Safe & Sub-Pixel Image Reads (#62)
Add ability to read images with sub-pixel coordinates (if upsampling while reading) and also to safely read areas outside the image (similar to OpenSlide read_region) without an index error. By default, it is zero padded i.e. black but can be changed to reflection, symmetric or other padding easily.

This PR additionally adds a small utility for finding the size of a bounds tuple: `bounds2size`.

The added functionality is split into two parts:

`safe_padded_read`
--------------------------

This allows reads outside the image region (0 - width and 0-height). It takes integer bounds but they may be any integers (negative or outside the image bounds).

![out_of_bounds_demo](https://user-images.githubusercontent.com/4615004/101831791-5d9c7080-3b2e-11eb-80a3-2ba3f1e84963.png)

`sub_pixel_read`
---------------------------------

This allows for sub-pixel reads. You can supply a set of bounds as floats and an output size. The region will be read using the fractional bounds and upscaled using a given interpolation method (defaults to cubic).

### Modularity & Potential Use Beyond Numpy Images
Internally this method calls safe_padded_read. However, this can be swapped out for any method via the read_func kwarg. read_func is given takes integer bounds, although they may be outside the image bounds. This can easily be replaced e.g. with some small wrapper around OpenSlide read_region or glymur read. This makes it more flexible and able to be used in other contexts.

### Interpolation Padding
This also accounts for boundary effects resulting from interpolation. By default, padding is added to the image before interpolating and then removed after. This is because some interpolation methods such as cubic and lanczos have a region of support over a few pixels which could extend outside the final visible region. This can be disabled with a key word argument.

### Fancy Rounding
There is also some fancy rounding in there to ensure the output image size is consistent with the requested output size. This may also be of use or inspiration for other read functions to help ensure this consistency. E.g. there is often an 1-off error with glymur reads.

### Illustration of various options

![sub_pixel_demo](https://user-images.githubusercontent.com/4615004/101831816-64c37e80-3b2e-11eb-975c-3ae8de608435.png)


The difference that extra padding for interpolation makes is very subtle. Here is the comparison zoomed in:

![sub_pixel_demo_interp_padding](https://user-images.githubusercontent.com/4615004/101831828-69883280-3b2e-11eb-8274-bab000f450f4.png)

## Tests
Tests have added a couple of tests: one for known troublesome parameters and another fuzz test. The fuzz test uses 10,000 sets of random parameters to check that these functions do not throw errors and produce consistent results i.e. output image size matches requested size.

Co-authored-by: John Pocock <John-P>