ENH: Add functionality to read through mask in `patchextraction` (#120)
This PR is focused on `patchextraction` functionality to clean up some part, change the API, and adding new features to it.

The `SlidingWindowPatchExtraction` class will accept the `input_mask` argument to extract patches based on the regions that are highlighted in the `input_mask`. The `input_mask` can be a path to the desired mask mask, a numpy array, `VirtualWSIReader`, or `'auto'`. In case of `'auto'`, a tissue mask is automatically generated for the `input_image` using tiatoolbox `TissueMasker` functionality if the `input_image` is a real WSI, otherwise, no mask will be considered for patch extraction.

- Rewrite the `slidingwindow` patch extraction class to use `get_coordinates` static method for patch bounding box generation.
- Add mask-based patch extraction functionality based on the `filter_coordinates` static method.
- Add tests for coverage.

Co-authored by @mostafajahanifar & @vqdang