NEW: TIFFWSIReader (#116)
Add support for SVS and OME-TIFF using tifffile, imagecodecs, zarr. Other TIFF formats can be supported in the future and potentially move away from relying on OpenSlide.

The bulk of the work in adding support for TIFF format in tiatoolbox is extracting and normalising the metadata. This needs extensive testing (unit and battle testing). The implementation of this class is very similar to that of the Omnyx JP2 reader with glymur, just with more metadata parsing.

### Work Log

1. Add dependencies for tiffile, imagecodecs, zarr.
2. Add TIFFWSIReader class.
3. Bug fixing (in metadata parsing for SVS and zarr grouping).

Co-authored-by: @John-P