ENH: Improve Mask-Based Patch Extraction Speed & Add Units Conversion (#217)
- Improve performance of mask based patch extraction.
In the current implementation, the `filter_coordinates` function is used to filter out generated coordinates based on the input mask. This is done by reading the mask region from `mask_reader` object (which is a VirtualWSI) at the same resolution and region requested in the patch extraction. Therefore, this process can take very long specially if the whole slide mask region is big and user requests to extract small patches from highest resolution (not to mention the increase in time due to overlap in patch extraction). For example, if you want extract 256x256 patches at 40x from the below WSI, it might take up to **10 minutes** for checking filter the coordinates

![output](https://user-images.githubusercontent.com/74412979/145984288-2de12b5a-6520-40ea-a41c-57bce1edda40.png)
![output_](https://user-images.githubusercontent.com/74412979/145984540-53edcce2-f4ec-41dc-bd9d-01febe73328b.png)

In this PR, I have introduced `filter_coordinates_fast` which uses the overview of `mask_reader` object (a numpy array) at a fixed-low resolution (mpp=8 or power=1.25) to speed of the region mask reading process and hence improve the speed of coordinate filtering and whole patch extraction processes. For example, the above image takes about **0.5 seconds** (at least **1,000x improvement in speed** in comparison to previous method).

![output__](https://user-images.githubusercontent.com/74412979/145985923-6b5eccea-94fb-4626-89f4-262644f8f314.png)

To be able to do this, I also introduce a method to `WSIReader` class that converts resolution from one unit to all other units, because the `filter_coordinates_fast` expects its input mask resolution to be in `mpp`, `power`, or `baseline` units. 

- [x] Implement fast coordinate filtering function
- [x] Implement resolution conversion function in wsireader
- [x] Writing tests
- [x] Improve coverage

Co-authored-by: Mostafa Jahanifar <74412979+mostafajahanifar@users.noreply.github.com>
Co-authored-by: Dang Vu <24943262+vqdang@users.noreply.github.com>
Co-authored-by: John Pocock <John-P@users.noreply.github.com>