MAINT: Setting Up Conditional Tests Using "pytest skipif" (#258)
There are several tests that have been marked to be skipped when running pytest because there were no way to check if they are run on Travis. Now we have the facility for detecting runtime flag in place, we can automatically flag which test to run on which conditions. This PR deals with tests that are related to Deep Learning Models (require GPU and Travis runtime check).



Co-authored-by: Dang Vu <24943262+vqdang@users.noreply.github.com>
Co-authored-by: Simon Graham <20071401+simongraham@users.noreply.github.com>
Co-authored-by: Shan Raza <13048456+shaneahmed@users.noreply.github.com>