BUG: Fix AnnotationStore Compile Option Check (#337)
- Fix compile options for test_annotation_stores.py

This has been failing due to a change in the compile flags of recent versions of sqlite3. JSON functions are now a core component of SQLite, not an extension that must be enabled via a compile option. Because of this the ENABLE_JSON1 flag was removed and OMIT_JSON flag added. 

For more info see: https://sqlite.org/src/doc/json-enhancements/doc/json-enhancements.md

The fix just adds a version check, and checks the compile options appropriately to the version.


Co-authored-by: John Pocock <John-P@users.noreply.github.com>