DEV: Add GitHub Action To Perform Conda Solve (#398)
Adds a GitHub action which will attempt to solve the conda env for Ubuntu and Windows. It will run:
- When a requirements*.yml file is pushed.
- On the 1st of the month at midnight (on develop).

It will time out if the environment is not solved in 10 minutes.

Windows conda currently takes longer than 10 minutes and so the timeout may need to be increased, or there may be some constraints which need adjusting there.