DEV: Add Pre-Commit Hook To Check Requirements Consistency (#392)
This adds a ~test~ pre-commit hook which checks that requirements files are consistent. It currently checks several criteria:

That for main/dev pairs (e.g. `requirements.txt` and `requirements_dev.txt`):
1. Any key present in main MUST also be in dev.
2. The version in main MUST match the version in dev.
3. The constraint in main MUST match the constraint dev.

For all requirements files (pip and conda), if the key is in AT LEAST TWO files:
1. The version MUST match across files with the requirement.
2. The constraint MUST match across files with the requirement.

It also prints out helpful errors such as:

```
AssertionError: shapely has different constraints: None (requirements.conda.yml) vs >=
1.8.1 (requirements.dev.conda.yml).
```

```
AssertionError: imagecodecs has inconsistent constraints: >=2021.11.20 (requirements.t
xt), >=2021.11.20 (requirements_dev.txt), >2021.11.20 (requirements.conda.yml), >2021.
11.20 (requirements.dev.conda.yml).
```

I have also updated the constraints for the files based on these checks in this PR.
I would like to turn this into a pre-commit hook (in addition to this or instead of?), but this is more of a proof of concept and feedback PR.

There is also a special case for setup.py. It is loaded differently but otherwise treated like pip requirements file.

Co-authored-by: David Epstein <22086916+DavidBAEpstein@users.noreply.github.com>