DEV: Missing Imports Check (Pre-Commit Hook) (#408)
### Found Imports Missing From Requirements

- [ x] `typing_extensions`: Needed for 3.7 support only, recommended by standard library.
- [ x] `scipy`
- [ x] `IPython`
- [ x] `flask`: Missing from `setup.py`, may be done on another branch already.
- [ x] `pkg_resources`: Part of `setuptools`, used for loading packaged data. There may be a better way to do this.


### More details

Some packages have a different package name to install from, PyPI versus the name used for importing (e.g. PIL, scikit-learn, etc.). Currently, this method assumed the name is the same, unless it is a known alias in a dictionary. This is to avoid having to send requests to PyPI to check. The PyPI servers can be slow and not always available. Perhaps this could be added with the dictionary as a fallback.

EDIT: It may be worth moving this into its own repo with some basic configuration (e.g. source root path, import aliased etc.) to make it a generic pre-commit hook that other repos can use also.

Co-authored-by: Adam Shephard <39619155+adamshephard@users.noreply.github.com>
Co-authored-by: Shan Raza <13048456+shaneahmed@users.noreply.github.com>