Improve 09-multi-task-segmentation.ipynb (#479)
The file 09-multi-task-segmentation.ipynb in branch doc-09 is the result of an overnight run on my Macbook, with ON_GPU=False. After the run, I changed this to the default ON_GPU=True. 

The colour bar shows the background as black. However, I think the background looks grey, so I think this may be a bug, but I'm not sure. The colour bar is produced by tiatoolbox.visualization.overlay_prediction_mask, which has been worked on recently by @Srijay-lab and, perhaps, by @measty.