:bug: BUG: Fix Level When Using DICOM `read_region` (#564)
This PR fixes the dicom slide thunbnail issue (maybe). the issue seems to arise from a difference in how the wsidicom reader treats levels, in that it labels them by their relative magnification (in power of 2).

so for example the levels in the wsidicom reader are [0, 2, 4] if the level downsamples are [1, 4, 16]

whereas in our WSIReader level is just an index [0, 1, 2], and the level_downsamples are separate.

My knowledge of DICOM is non-existent, so it may be that the fix is wrong - hopefully john can have a look at it. It fixed the thumbnail issue for me though.

----------------------------------------

Co-authored-by: Mark Eastwood <20169086+measty@users.noreply.github.com>
Co-authored-by: John Pocock <John-P@users.noreply.github.com>
Co-authored-by: Shan Raza <13048456+shaneahmed@users.noreply.github.com>