🐛Fix `test` in `test_annotation_stores.py` (#779)
Can no longer pass None to `pytest.warns()`. This is why tests were failing recently in test_annotation_stores.py. See: https://docs.pytest.org/en/latest/deprecations.html#using-pytest-warns-none

They seem to have left no convenient way to test for no warnings using pytest.warns(), so I have rewritten that test to explicitly check that the warning doesn't exist in the log.

Unfortunately, tests are also failing in test_stainnorm.py due to an error:
```
FAILED tests/test_stainnorm.py::test_cli_stainnorm_dir - AssertionError: assert 1 == 0
 where 1 = <Result LinAlgError('SVD did not converge in Linear Least Squares')>.exit_code
```

I have no idea why this is happening and cannot replicate it on my machine, it may be related to the age old problem of vahadane normalization being flaky as that is the thing it is doing in that test. This is backed up by the fact that changing the type of stain normalization done in the cli stainnorm test to ruifrok seems to remove the issue.

It is odd that vahadane SN is passing tests when run programatically, but is having this issue when used as part of a command line test. We will probably need to look deeper into what is going on here.