🐛 Fix `json_extract` Construction (#772)
Restructures the way the dsl builds the SQL expression to access JSON properties so that properties with a . in are correctly treated as a single key

Adds a test to check for correct behaviour when accessing a property containing a .

Addresses issue #771 

Would like to check with John if removing `json.dumps()` is ok. I couldn't see a reason it was necessary and all tests pass/all the use-cases I tried worked fine, but maybe there is one I missed.